Split colour extensions out of UIExtensions into Extensions group
