+++ CodeWordBreaker/CodeWordBreaker/UI/Extensions/UIColor+Hex.swift
@@ -1,0 +1,16 @@
+//
+//  UIColor+Hex.swift
+//  CodeWordBreaker
+//
+//  Created by Alexander Ostrovsky on 3/4/2026.
+//
+
+import UIKit
+
+extension UIColor {
+    var hex: String {
+        var r: CGFloat = 0, g: CGFloat = 0, b: CGFloat = 0, a: CGFloat = 0
+        getRed(&r, green: &g, blue: &b, alpha: &a)
+        return String(format: "#%02X%02X%02X", Int((r * 255).rounded()), Int((g * 255).rounded()), Int((b * 255).rounded()))
+    }
+}
