+++ CodeWordBreaker/CodeWordBreaker/UI/Extensions/UIExtensions.swift
@@ -48,40 +48,6 @@ struct ShakeEffect: GeometryEffect {
     }
 }
 
-extension Color {
-    static func gray(_ brightness: CGFloat) -> Color {
-        .init(hue: 148/360, saturation: 0, brightness: brightness)
-    }
-}
-
-extension Color {
-    init(_ hex: UInt, alpha: Double = 1) {
-        self.init(
-            .sRGB,
-            red: Double((hex >> 16) & 0xFF) / 255,
-            green: Double((hex >> 8) & 0xFF) / 255,
-            blue: Double(hex & 0xFF) / 255,
-            opacity: alpha
-        )
-    }
-
-    init?(hex: String) {
-        guard hex.hasPrefix("#"), hex.count == 7,
-              let value = UInt(hex.dropFirst(), radix: 16) else { return nil }
-        self.init(value)
-    }
-
-    var hex: String { UIColor(self).hex }
-}
-
-extension UIColor {
-    var hex: String {
-        var r: CGFloat = 0, g: CGFloat = 0, b: CGFloat = 0, a: CGFloat = 0
-        getRed(&r, green: &g, blue: &b, alpha: &a)
-        return String(format: "#%02X%02X%02X", Int((r * 255).rounded()), Int((g * 255).rounded()), Int((b * 255).rounded()))
-    }
-}
-
 extension View {
     func trackElapsedTime(in game: CodeWordBreaker) -> some View {
         modifier(ElapsedTimeTracker(game: game))
